// vite-project/src/interface/client/login.ts
export interface SignInInterface {
    email: string;
    password: string;
}
export interface CheckValidInterface {
    email?: boolean;
    password?: boolean;
    title?: boolean;
    price?: boolean;
    brand?: boolean;
    category?: boolean;
    image?: boolean;
    ram?: boolean;
    color?: boolean;
    capacity?: boolean;
    description?: boolean;
    seller?: boolean;
    quantity?: boolean;
    valueAddNewBrand?: boolean;
    blog?: boolean;
}